Use FFI instead of Rust types for constants too
Otherwise we will make them to possibly incorrect Rust types, which do
not exist, and by this we map them to the same types as used inside
structs, function parameters, etc.

Case in question was GStreamer's GST_CLOCK_TIME_NONE which is of type
GstClockTime in C. By using the Rust type this would map to "ClockTime",
which does not exist in the FFI bindings.