Generate *Ext traits by default and allow disabling them instead
Reason for this is that the general case is that subclasses of any
GObject are allowed, and in your own crate you don't generally know if
other crates or any other code is going to create subclasses. Only in
few cases a class is "final" and does not allow subclasses.