Implement Send/Sync configuration for objects/records
This allows configuring crate-wide and per object/record if an object is
supposed to be Send, Sync or Send + Sync. The marker traits will
automatically be implemented and signals callbacks on those objects will
get the appropriate trait bounds.

Note that it is still required to manually ensure that any signal of
parent or child types has at least the same trait bounds.

https://github.com/gtk-rs/gir/issues/379