Ignore action signals
These work the other way around than normal signals: We are supposed to
emit them from outside the object instead of connecting to it by using
g_signal_emit*(). It basically is like a dynamic function call.
The object itself is connected to the signal and waiting for us to emit
it, to do something.