Generate enums in -sys crates as integer constants
The C code might be extended by additional values without the bindings
being updated, which then is considered undefined behaviour on the Rust
side. As such we must use plain integers here.

The enums stay as enums in the non-sys crates and the conversion traits
ensure that the conversion from the integers to the actual enum values
is done correctly, while storing unknown values in a hidden enum
variant.

Fixes https://github.com/gtk-rs/gir/issues/480