Migrate to `tempfile` from deprecated `temdir`
The `tempdir` crate was deprecated by its authors, and `tempfile` is
recommended as a replacement.

I assume this is the first time that a dependency needs to be removed
from `Cargo.toml` files, because I found no functionality to remove the
entry. I created an `unset` function analogous to the `set_string`
function found in the `cargo_toml.rs` file.

Rationale: I do packaging work for rust crates in Debian, and due to
it's deprecation, we decided to not package the `tempdir` crate at all,
but to submit patches to the projects that still use it instead.