Correctly get function configuration and make use of it for get_type() functions in sys mode
We shouldn't look for the full get_type function name but simply for
"get_type". As there is only the associated C function name we simply
define "get_type" as the name to look for in the configuration.

Once we have that, make use of the ignore and version configuration in
addition to the general #[cfg] handling. This allows us to ignore
get_type() functions and override their version, for example if the
get_type() function was added after the type like in Pango.

First part of https://github.com/gtk-rs/gir/issues/839