Use write_str instead of write_fmt when no formatting is needed
Write_fmt (what write! expands to) needs to be set up with an
std::fmt::Arguments object that is consequently iterated.  All segments
of the format string get Display::fmt called on them and those
implementations usually result in calling some form of write_str (at
least for &str).

This is a minor change that only saves few unnecessary cyles and is
mostly done for cosmetic reasons.  It finishes the work started in
3eaed54.