codegen/function: Omit version constraints if already guarded by scope
This removes unnecessary (duplicate) `#[cfg]` attributes on `fn`s where
the surrounding `impl` block (in `enums`/`flags`) or `mod` (when in a
file that fully defines a `struct`) already declares this constraint.
When the feature requirement is already described it is superfluous (and
noisy) to specify it again.  Besides, rustdoc omits duplicate feature
requirements if it can show this just once at the top of the page for
the entire `struct`.