Remove use_boxed_functions override
This override only works when get_type is available and is intended when
copy and free are not available, to call the generic g_boxed_copy/free
functions (which under the hood resolve the right copy/free function
from its GType, as retrieved with get_type).

However, this behaviour is already implied if copy/free nor (un)ref are
found. Hence it is only useful to forcibly use generic g_boxed_copy/free
even if type-specific functions are directly available in G-IR and Rust.
That can almost never be the right thing.

The functionality isn't used for nearly 3 years, and wasn't even used
for all that long in gstreamer-rs; it was introduced in [1] and removed
a mere 15 days after [2].

[1]: https://gitlab.freedesktop.org/gstreamer/gstreamer-rs/-/commit/3a0c9723044fb3c9063fcef60e0270171576ba92
[2]: https://gitlab.freedesktop.org/gstreamer/gstreamer-rs/-/commit/92016945fc7882979664bbea6e36a5585a2fea1d