function: Teach gir how to generate whole functions as unsafe
While implementing GLFilter bindings review suggested to mark a few new
functions as unsafe.  It saddened me to have to copy these functions to
"manual" files and repeatedly update/maintain them as the main repo and
GLFilter MR progressed: notable are the addition of doc(alias) and
mutability changes.  Not to mention changes to versioning cfg() blocks
that those functions were luckily not affected by.