analysis/special_functions: Import glib::translate for mut_override
analysis/functions was cleaned up in the commit before to not emit
`glib::translate::*` imports for special functions; those are
responsible to provide the right imports themselves. Fortunately `copy`
was recently improved to specify whether it needs `mut_override`, which
can now also be used to conditionally provide this import.