Added support for connecting to signals with details
Some GObject signals have a non-default _detail_ argument [1]. This
allows one to specify that a closure should only be invoked on a
subset of that signal's emissions. With this commit, Gir will add a
'detail' argument (an Option<&str>, as being invoked on all possible
values of _detail_ is allowed and sometimes desired) to the
connect_<signal_name> methods generated for signals whose metadata
notes that non-empty values of _detail_ get used.

This is a squashed version of the history for pull request
gtk-rs/gir#1062.

[1] https://developer.gnome.org/gobject/stable/signal.html#signal-detail

Co-authored-by: Marijn Suijten <marijns95@gmail.com>