git: Read remote url from upstream or origin
With the new separate gir directories it is useful to know what repo
(url) a git hash came from, without hardcoding it here. Consensus shows
that these repos are either directly cloned through submodules with a
default remote named `origin` or a manual rename to `upstream` with the
personal remote (from which PRs are commonly created) named `origin`. In
that case the url of `upstream` is preferred.