git: Read URL for submodules from .gitmodules
As an improvement over the previous commit, which deals with custom
setups where contributors and maintainers have multiple remotes set up
under a specific name, this approach reads the URL from the .gitmodules
file directly. This file is generally not updated as that would cause
the tree to become dirty while still allowing local modifications
(changing and renaming remotes) to the checkout itself.