doc: Use intra-doc-links instead of relative paths to html pages
Using intra-doc-links saves quite a bit of code in gir, looks cleaner in
the markdown files and resulting docs (`[src]` button mostly), and has
the advantage of being checked by `rustdoc` for validity.

That last point is not totally fair as most of these trait "implements"
/ "implementors" segments are generated inside macros, for which
`rustdoc` does not seem to emit any warnings or errors currently.