codegen/doc: Omit "Feature: `vXXX`" text in favour of doc_cfg
Nowadays we have the `doc_cfg` feature rendering clean, consistent and
obvious markers for feature requirements through `doc(cfg())` attributes
in the code (perhaps automatically derived from `cfg()` attributes in
the not so distant future). With this it is not necessary to repeat the
requirement in the documentation text once again, which wasn't really
obvious nor good-looking to begin with.

Besides, this "Feature:" text is currently erroneously generated on the
last member of an enum, instead of on the enum itself (parts omitted for
brevity):

    #[cfg(any(feature = "v1_14_1", feature = "dox"))]
    #[cfg_attr(feature = "dox", doc(cfg(feature = "v1_14_1")))]
    pub enum WebRTCFECType {
        /// none
        None,
        /// ulpfec + red
        ///
        /// Feature: `v1_14_1`
        ///
        UlpRed,
    }