codegen/doc: Only emit Self:: for @-references if parent type matches
`@` is used in few invalid places in the (GStreamer) docs where `#`
should have been used.  Instead of generating broken links to `Self::`
here generate a `crate::` link that at least still works, with an
obvious comment for future improvement in the C codebase.

If this is fixed the assert can finally be flipped on.