codegen/doc: Generate intra-doc-links for Self:: (`@`) references
This is using an explicit reference to `Self::` so that there is no
confusion over what to reference.  Enum variants with `None` would
otherwise reference Rusts `None` primitive.  `Symbol` neatly already
contains the symbol name for us, in addition to the owner and other
fields that would otherwise form a full Rust name.

Note that there is no validation whether Symbol::owner_name actually
equals the type the documentation is being generated for; this is down
to the C documentation.