codegen/doc: Add `()` to function symbols
This already exists in the C docs and is clearer to the reader that a
function is referenced and not something else.  It becomes particularly
relevant when generating intra-doc-links, where rustdoc will emit
clearer error messages when it knows the target is supposed to be a
function but is not or doesn't exist at all.