codegen/doc: Wrap virtual functions properly and use intra-doc-links
Virtual functions ar generally referred to as #Type.vfunc() which would
get treated as `Type.vfunc`().  By checking for (variants of) this
pattern first the `()` can be incorporated in the markdown syntax and
links resulting in [`Type.vfunc()`].  Extra checks are used to place
SElf:: or crate:: on the type where necessary or more elegant.