doc/format: Emit intra-doc-links for symbol references
As final step in the intra-doc link conversion this enables generating
links for `#` and `%` references. These references are tricky because
they may not all exist in Rust (ie. memory cleanup functions are taken
care of by drop handlers) resulting in lots of warnings from rustdoc.

Similar to previous "self" (`@`) and function links this takes care of
using simplified `Self::` links where possible, and otherwise prepends
`crate::` to make sure the type or symbol is found in global scope.