config: Normalize submodule path before looking it up in .gitmodules
`normalize_path` was introduced in [1] for pretty-printing relative
paths with `../` ParentDir components.  It was not intended for use
outside of displaying scenarios but thanks to tests we can safely use it
to solve a mismatch [2] when searching for submodule paths in
`.gitmodules`.

This functionality introduced with [3] finds the submodule associated
with a path based on string matching in order to pull out the URL, but
was only tested with GStreamer-rs where "clean" paths are passed in
through gir's `-d` option.  Gtk-rs and gtk4-rs use `girs_directories =`
in `Gir.toml` with relative paths like `gtk4/../gir-files`: these cannot
be found in `.gitmodules` with simple string matching and need the path
simplified first.

[1]: https://github.com/gtk-rs/gir/pull/1075
[2]: https://github.com/gtk-rs/gtk4-rs/pull/251#discussion_r606147518
[3]: https://github.com/gtk-rs/gir/pull/1068