Support types implementing try_from_glib
Types implementing `TryFromGlib` can now be configured to define
how code generation should handle the resulting type. The new
`TryFromGlib` `struct` contains information on the type resulting
from the configuration of the type itself. During analysis, a
parameter or return value can override whether it should be
considered `mandatory` or `infaillible`. This requires the
analyzed data to be taken into account when building the
parameter or return value's type.

As a consequence, the `TryFromGlib` instance resulting from
analysis must be passed to the `rust_type` functions.

When the resulting type is `Result<OkType, ErrType>`, two types
need to be used. This was not possible with previous model since
the result from `used_rust_type` would contain only one `String`.
The new `struct` `RustType` strives to maintain the same API
regarding the result values for the `rust_type` functions while
allowing multiple `use`d types. This renders the `used_rust_type`
uselees as everything can now be handled by the `rust_type`
functions, which also allows for a single call in some places.