config,git: Omit gir directory hash and URL if not a submodule
Intially plaguing GStreamer-rs this code grabs the git hash from the
root repository if the specified gir directory is not a submodule.
GStreamer-rs since moved all gir files to subdirectories (because it is
possible to have multiple now) but various external projects have a
single gir file checked in to the root repo and do not want to see all
generated files change the hash after every regeneration while the
root repository is checked out on a different commit.

This change uses `git rev-parse --show-toplevel` to resolve where the
repository (folder containing `.git`) for a given directory lives and
ignores the hash and URL if that is equivalent to the canonical path of
the root.