analysis/record: Do not bail on missing memory functions
This constraint is already enforced with a `panic!` during codegen,
effectively disallowing Gir.toml from specifying objects that cannot be
generated.  It is now more obvious instead of being just an error line
in the (already pretty spammy) logs.

Furthermore analysis is also used to generate documentation, including
those for manual types.  Bailing here means no analysis for the type
is available leading to missing documentation.  After this change we get
a bunch of missing text to show up in `docs.md` in GStreamer for the
manual types that lack memory functions.

See also #1094