generator: Embed documentation using rustdoc-stripper after generating
We have decided to get rid of checked-in documentation in favour of
performing generation and embedding at once in the CI (or locally for
testing) in a single script.  Since the generator is already looping
over crates containing Gir.toml this is the only natural place for that.

This change removes the custom documentation output path again (leaving
it to the default `<crate>/docs.md`) and invokes `rustdoc-stripper -g`
straight after to embed the documentation.  The command-line flag is now
a simple boolean to enable this mode; without it Rust files are
generated as normal.