codegen/object: Only re-export traits from `traits`/`prelude` module
Following a discussion in [1] we have decided to remove all trait
exports - both manual and automatic - from the crate root in favour of
having them available to crate users in `prelude` only, as per the
intended design.  It was inconsistent to have ExtManual traits available
in `prelude` only (and some randomly in the crate root in GStreamer-rs)
whereas all automatic traits were re-exported through the crate root as
well.

[1]: https://github.com/gtk-rs/gir/pull/1111