build: Rerun if git hash changed
`gir`'s hash in version files gets out of sync quite frequently with the
hash of the submodule.  This happens when generating from a change
locally, creating a PR and later regenerating on the merge commit.  When
no file changes have been made `gir` doesn't get recompiled with the
newer hash, but the submodule hash is still changed resulting in a
mismatch.  It is usually caught later on another machine/checkout.

This change ensures `gir`'s hash is always updated when the checkout
changes, even if no code changes have been made that would normally
cause `cargo` to recompile it.