codegen/enums: Keep emitting name of `Self` in string literals
Commit 3f36d06 ("Fix some clippy::use_self warnings in generated code")
did away with repeating the surrounding type name in places where the
`Self` keyword is preferred (`clippy::use_self`), but accidentally also
replaced the type in a string literal used for formatting (first noticed
in [1]).  When these enums are printed the logs will be filled with
helpless `Self::<variant>` instead of `<type>::<variant>` which wasn't
intended.

[1]: https://github.com/gtk-rs/gtk4-rs/pull/349/files#r625049878