codegen: Emit `crate::` for all global `%` constants except from prelude
Rust prelude symbols - `true`, `false` and `None` - are explicitly added
to the list of symbols.  These are the only global constants/symbols
that shouldn't receive a `crate::` prefix.  Everything else coming from
the current or another crate must be imported from the root as it may
reside in another file/module.