codegen/doc: Drop `[Deprecated]` text in favour of Rust annotations
All deprecated objects, traits and records already receive a
`#[deprecated = "Since XXX"]` annotation in Rust which neatly shows up
in the docs.  The extra `[Deprecated]` label in the documentation is not
necessary anymore.

`# Deprecated` headers remain for the documentation part clarifying why
an item has been deprecated, and what its alternative is.