codegen/doc: Use `[][]` for intradoc links instead of `[]()`
Rustdoc treats `[foo]` as item path, and by the same rule
`[fancy text][foo]` too.  Unlike `[]()`, which generates a dead link if
the right-hand side is a valid Rust item path but doesn't exist in scope
(common and unfortunately unavoidable with docs for C code auto-ported
to Rust), the `[]` and `[][]` do not generate a link at all.