analysis/bounds: Panic if all type parameters are exhausted
Most of the callers were not reading the return value leading to
potentially missed failed type parameter "allocations".  This could have
been solved with `#[must_use]` but the function might as well panic
directly as this situation is extremely unlikely (need more than `11`
type parameters) but nevertheless not valid.  If it occurs in the
future `gir` should be adjusted to have more type parameters available,
ie. by using full names.