anlysis: Ignore per-type `ref_mode` override for input transfer-full
This override is mainly used to turn common mutability issues for all
parameters of a specific type down to immutable borrows when the
underlying C source cannot be patched as it uses interior mutability for
refcounting etc.
The override is only used in GStreamer, but is unfortunately applied in
such a way that the few cases using `transfer-full` parameters also get
their argument turned into a borrow, leading to incorrect API and
sometimes failed compilation when trait bounds can (logically) not be
satisfied.  This is simply addressed by returning `RefMode::None`
whenever an override is provided for an input parameter that happens to
be `transfer-full`.