enums,flags: Always analyze manual types
Docs generation is going to look at analysis data to know what variants
and functions are available on enums/flags.  External types (in the
`manual = []` list) were previously omitted because of their status.
This changes enums and flags to behave like objects and records which
only bail if the type is entirely ignored, and perform a
`need_generate()` check at `codegen` instead.

Fixes: 50cee9e ("enums: Move rejection and glib imports from codegen to analysis")
Fixes: e549975 ("Copypaste enum analysis and function generation to bitfield")