analysis/function: Inherit GStatus from owning object
When the surrounding type is ignored or manually implemented the
underlying functions are never generated; make their `status` field
reflect this case.

This fixes a future case where flags and enums are going to be analyzed
to have information available in documentation generation, despite being
marked `manual`.  These flags and enums collect all their imports in one
object which have to be omitted for `manual` types, and an easy way to
do that for the functions on types is to take advantage of the
`status.need_generate()` check before function analysis starts analyzing
and adding imports.