Use or_patterns, stabilized in Rust 1.53
Previously `|` was only allowed in the top-level of patterns, requiring
the circumfering enum variant (usually something wrapped in `Some`, and
in `gir` `Type::Fundamental` is common) to be repeated for every inner
variant that's matched against.  Not anymore, since 1.53 it is now
possible to write `|` within a single variant instead of repeating it
many times.