Remove the `links` annotations
The `links` annotation in `Cargo.toml` is intended to ensure that in the
crate graph there's at most one crate that's an implementation of
some sort concept.

This can make sense in some scenarios, most prominent of which is when
the crate defines `#[no_mangle]` symbols (e.g. by compiling a vendored C
library.) In that situation linking a binary that depends on two
versions of the library cannot work because of colliding symbol names.

There does not appear to be a similar reason to impose such a
restriction on the users of gir-generated bindings, however. All of
the generated crates link to a system library, they do not define any
`#[no_mangle]` symbols nor they vendor/build C libraries as part of
their build process.  Most likely all the different versions of the
bindings will link to the exact same system library too.