build: Emit rerun-if-changed for branch file on `ref: ` HEAD
We already ran into the issue of git hashes not updating when checked
out to a branch, because the `HEAD` file doesn't change.  `HEAD` does
however tell us what branch (ref) the user is checked out to, which is a
file much like `HEAD` containing the hash of the commit it is currently
pointing to, and will inevitably change if the user "moves around" (ie.
rebasing or committing).  Emitting a `rerun-if-changed` on this file
should make the git hash much more reliable.