Remove type version heuristics
Previously we would infer the minimum version of a type by the minimum
version of all its functions if not type version information was
available. This causes unnecessarily high minimum version constraints if
the type was available without functions before and later got functions.

Instead this should be fixed in the C library's annotations, and for now
this can also be overridden via Gir.toml if needed.