codgen: don't append () to gi-docgen not found functions
Otherwise, the gtk-doc style parsing code, will detect those as functions which ends up breaking
the whole docs, such as issues seen in #1200. Ideally, we would have more robust code
that detects whether the detected item is already a rust item or an item we couldn't found the first time with gi-docgen
to avoid breaking things. But such features are not possible with the currently use regex rust crate