codegen/special_functions: Add missing space between `pub` and `fn`
Originally the `visibility` modifier contained this space ("pub ") to
keep indentation in check.  A recent change however formats the
visibility modifier from configuration directly into the file, without
inserting the necessary trailing space, resulting in the text `pubfn` in
`special_functions`, which is invalid syntax.

Since output is almost always formatted with rustfmt anyway, this
indentation mismatch is irrelevant to allocate a string for here.

Fixes: e3bd76a3 ("replace FunctionVisibility with hidden/commented attributes")