analysis: suppress assertion for throw function to return void
In GNOME convention, the function categorized to throw return boolean
value for the result of function call. When it's true, the argument of
the function for GError is set. However it's not necessarily followed by
actual libraries. Some libraries has function to return void with
argument for GError. They expect callers to detect error by checking the
value of argument for GError.

This commit relaxes restriction from the convension. The assertion is
not inserted to check return value for throw function.