git: add non unix fallback for creating PathBuf from Vec<u8>
We obviously can't use ::unix:: on non unix platforms.

Not sure the fallback will work well on all non unix platforms.
Time will tell...

See https://users.rust-lang.org/t/is-there-a-way-to-convert-vec-u8-to-osstring/49970
and https://users.rust-lang.org/t/stdin-stdout-stderr-and-encoding/5354
Not sure if any of this applies to rust std::process::Command.

Please note the use, in the same file, of a more platform agnostic approach when  a String is needed.
For example:
    let hash = String::from_utf8(output.stdout).ok()?;

fixes https://github.com/gtk-rs/gir/issues/1360