book: Unbreak syntax highlighting on `console` codeblocks, address concerns consistently
[#1365] was blatantly merged without considering review, breaking syntax
highlighting on these codeblocks in the process.  As suggested, when
removing the "effective UID" prefix `console` codeblocks loose their
formatting and must instead be converted to regular `sh` or `bash`
types.

At the same time, make this change consistently across the book, instead
of considering `sys_library.md` to be the only page.  Note that it has
been retained in places where a `console` block is intented, because the
lines after a console command represent program output and are formatted
accordingly when no "effective UID" prefix is detected.

[#1365]: https://github.com/gtk-rs/gir/pull/1365