generator.py: Restore pre-building `gir` with build/progress output
In commit c6ba79ed ("generator.py: Use cargo run instead of guessing
path") the `cargo build` command on the default (hardcoded) `"gir"`
directory was replaced with a single `cargo run` for correctly finding
the right executable.

This isn't an issue given that `cargo run` also builds the tool, and
above all makes it easier / more correct to find the binary as the
primary reason for that commit, but it does make us lose the build log
as a progress indicator (thanks to `--quiet` and capturing `stderr`).
In addition, it causes the first `` => Looking in path `xxx` `` item
to take a very long time without aforementioned progres indication,
making the user wonder what is going on.

By building `gir` again with a separate `cargo build` invocation before
running, this issue is alleviated somewhat.