Generalize Bound & to_glib_extra in more contexts
This commit paves the way for the introduction of additional bounds in more
contexts by generalizing the use of `Bound` and `to_glib_extra`.

`to_glib_extra`s allow prepending a processing before calling the actual
conversion to the glib type. For instance, this can be used to convert a value
into an `Option` when an `impl Into<Option<_>>` argument is used.

This commit adds support for `to_glib_extra` for callback arguments and for
property setters.

Properties were using a specific `PropertyBound` type which added code
duplication and couldn't take advantage of the regular `Bound` which can take
advantage of `to_glib_extra`.