Don't skip `Option` for nullable collections where applicable
Nullable collection were represented as a `Vec<_>` or `[_]`, but an empty
collection is not the same as an undefined collection.

Don't do this for return and output values though: in practice the C
implementations return NULL when the size is 0. So it doesn't make much
sense using an `Option<_>` in that case.

Finally, Gir now parses the `zero-terminated` annotation which allows
detecting functions returning an array for which the length is not defined
by a length argument. Those functions used to require an explicit `manual`
declaration in the `Gir.toml` and are automatically commented out now.

Fixes: https://github.com/gtk-rs/gir/issues/1133
See also: https://github.com/gtk-rs/gtk-rs-core/issues/1257