codegen: Take ns into account for deprecations
Suppose we have two libraries A and B and an object in B is a subclass of an object in A.

Given that B will generate {g,s}etters for properties in A, if any
property in A gets deprecated, then the namespace would be ignored when
regenerating the bindings.

One example is the Gtk.Application:register-session property in GTK
which was deprecated in 4.22. In adw::Application we would land with

```
    #[cfg_attr(feature = "v4_22", deprecated = "Since 4.22")]
```

instead of

```
    #[cfg_attr(feature = "gtk_v4_22", deprecated = "Since 4.22")]
```

This commits fixes this issue.